Move the SseResponseStreamTransport out of sample and into the main project
