+++ samples/AspNetCoreSseServer/AspNetCoreSseServer.csproj
@@ -10,8 +10,4 @@
     <ProjectReference Include="..\..\src\ModelContextProtocol\ModelContextProtocol.csproj" />
   </ItemGroup>
 
-  <ItemGroup>
-    <PackageReference Include="System.Net.ServerSentEvents" VersionOverride="10.0.0-preview.1.25080.5" />
-  </ItemGroup>
-
 </Project>
