+++ src/ModelContextProtocol/Protocol/Transport/SseResponseStreamTransport.cs
@@ -3,12 +3,15 @@ using System.Net.ServerSentEvents;
 using System.Text.Json;
 using System.Threading.Channels;
 using ModelContextProtocol.Protocol.Messages;
-using ModelContextProtocol.Protocol.Transport;
 using ModelContextProtocol.Utils.Json;
 
-namespace AspNetCoreSseServer;
+namespace ModelContextProtocol.Protocol.Transport;
 
-public class SseServerStreamTransport(Stream sseResponseStream) : ITransport
+/// <summary>
+/// Implements the MCP SSE server transport protocol using the SSE response <see cref="Stream"/>.
+/// </summary>
+/// <param name="sseResponseStream">The stream to write the SSE response body to.</param>
+public sealed class SseResponseStreamTransport(Stream sseResponseStream) : ITransport
 {
     private readonly Channel<IJsonRpcMessage> _incomingChannel = CreateSingleItemChannel<IJsonRpcMessage>();
     private readonly Channel<SseItem<IJsonRpcMessage?>> _outgoingSseChannel = CreateSingleItemChannel<SseItem<IJsonRpcMessage?>>();
@@ -16,8 +19,15 @@ public class SseServerStreamTransport(Stream sseResponseStream) : ITransport
     private Task? _sseWriteTask;
     private Utf8JsonWriter? _jsonWriter;
 
+    /// <inherityydoc/>
     public bool IsConnected => _sseWriteTask?.IsCompleted == false;
 
+    /// <summary>
+    /// Starts the transport and writes the JSON-RPC messages sent via <see cref="SendMessageAsync(IJsonRpcMessage, CancellationToken)"/>
+    /// to the SSE response stream until cancelled or disposed.
+    /// </summary>
+    /// <param name="cancellationToken">A token to cancel writing to the SSE response stream.</param>
+    /// <returns>A task representing the send loop that writes JSON-RPC messages to the SSE response stream.</returns>
     public Task RunAsync(CancellationToken cancellationToken)
     {
         void WriteJsonRpcMessageToBuffer(SseItem<IJsonRpcMessage?> item, IBufferWriter<byte> writer)
@@ -28,7 +38,7 @@ public class SseServerStreamTransport(Stream sseResponseStream) : ITransport
                 return;
             }
 
-            JsonSerializer.Serialize(GetUtf8JsonWriter(writer), item.Data, McpJsonUtilities.DefaultOptions);
+            JsonSerializer.Serialize(GetUtf8JsonWriter(writer), item.Data, McpJsonUtilities.DefaultOptions.GetTypeInfo<IJsonRpcMessage?>());
         }
 
         // The very first SSE event isn't really an IJsonRpcMessage, but there's no API to write a single item of a different type,
@@ -39,8 +49,10 @@ public class SseServerStreamTransport(Stream sseResponseStream) : ITransport
         return _sseWriteTask = SseFormatter.WriteAsync(sseItems, sseResponseStream, WriteJsonRpcMessageToBuffer, cancellationToken);
     }
 
+    /// <inheritdoc/>
     public ChannelReader<IJsonRpcMessage> MessageReader => _incomingChannel.Reader;
 
+    /// <inheritdoc/>
     public ValueTask DisposeAsync()
     {
         _incomingChannel.Writer.TryComplete();
@@ -48,14 +60,29 @@ public class SseServerStreamTransport(Stream sseResponseStream) : ITransport
         return new ValueTask(_sseWriteTask ?? Task.CompletedTask);
     }
 
-    public Task SendMessageAsync(IJsonRpcMessage message, CancellationToken cancellationToken = default) =>
-        _outgoingSseChannel.Writer.WriteAsync(new SseItem<IJsonRpcMessage?>(message), cancellationToken).AsTask();
+    /// <inheritdoc/>
+    public Task SendMessageAsync(IJsonRpcMessage message, CancellationToken cancellationToken = default)
+    {
+        if (_sseWriteTask is null)
+        {
+            throw new InvalidOperationException($"Transport is not connected. Make sure to call {nameof(RunAsync)} first.");
+        }
 
+        return _outgoingSseChannel.Writer.WriteAsync(new SseItem<IJsonRpcMessage?>(message), cancellationToken).AsTask();
+    }
+
+    /// <summary>
+    /// Handles incoming JSON-RPC messages received on the /message endpoint.
+    /// </summary>
+    /// <param name="message">The JSON-RPC message received.</param>
+    /// <param name="cancellationToken">A token to cancel the operation.</param>
+    /// <returns>A task representing the potentially asynchronous operation to buffer or process the JSON-RPC message.</returns>
+    /// <exception cref="InvalidOperationException">Thrown when there is an attempt to process a message before calling <see cref="RunAsync(CancellationToken)"/>.</exception>
     public Task OnMessageReceivedAsync(IJsonRpcMessage message, CancellationToken cancellationToken)
     {
-        if (!IsConnected)
+        if (_sseWriteTask is null)
         {
-            throw new McpTransportException("Transport is not connected");
+            throw new InvalidOperationException($"Transport is not connected. Make sure to call {nameof(RunAsync)} first.");
         }
 
         return _incomingChannel.Writer.WriteAsync(message, cancellationToken).AsTask();
